telegram_bot/pyproject.toml
[tool.poetry]
name = "radio-bot"
version = "0.1.0"
description = ""
authors = ["Ayaan Zaidi <hi@obviy.us>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
spotipy = "^2.23.0"
python-telegram-bot = { version = "^20.6", extras = [
  "rate-limiter",
  "job-queue",
] }
cerberus = "^1.3.5"
coloredlogs = "^15.0.1"
yt-dlp = "^2023.10.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
